# api-docs.json
{"openapi":"3.0.1","info":{"title":"REST API's RESTful by Blog Application with Java, Spring, Docker and Kubernetes","description":"REST API's RESTful by Blog Application with Java, Spring, Docker and Kubernetes","termsOfService":"https://joaojunio-dev.vercel.app","license":{"name":"Apache 2.0","url":"https://joaojunio-dev.vercel.app"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/auth/refresh/{username}":{"put":{"tags":["Auth"],"summary":"Updating token","description":"Method for updating the validation token.","operationId":"refreshToken","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{},"application/xml":{},"application/yaml":{}}},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"401":{"description":"Unauthorized"},"500":{"description":"Internal Server Error"}}}},"/api/post/v1":{"get":{"tags":["Post","Person"],"summary":"Finds all Post's","description":"Finds all Post's","operationId":"findAll","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostController"}}}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"put":{"tags":["Post"],"summary":"Update one Post","description":"Update one Post","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PostDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/PostDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"post":{"tags":["Post"],"summary":"Create Post","description":"Create Post","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PostDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/PostDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/person/v1":{"get":{"tags":["Person"],"summary":"Finds all people","description":"Finds all people","operationId":"findAll_1","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonController"}}}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"put":{"tags":["Person"],"summary":"Update one person","description":"Update one person","operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"post":{"tags":["Person"],"summary":"Create person","description":"Create person","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/like/v1":{"get":{"tags":["Like"],"summary":"Finds a Like's","description":"Finds a Like's","operationId":"findAll_2","responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"put":{"tags":["Like"],"summary":"Update a Like","description":"Update a Like","operationId":"update_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/LikeDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/LikeDTO"}}},"required":true},"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"post":{"tags":["Like"],"summary":"Create a new Like","description":"Create a new Like","operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/LikeDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/LikeDTO"}}},"required":true},"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/comment/v1":{"get":{"tags":["Comment"],"summary":"Finds all Comment's","description":"Find by one Comment's","operationId":"findAll_3","responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"put":{"tags":["Comment"],"summary":"Update a Comment","description":"Update a Comment","operationId":"update_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/CommentDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}},"required":true},"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"post":{"tags":["Comment"],"summary":"Create Comment","description":"Create Comment","operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/CommentDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/CommentDTO"}}},"required":true},"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/auth/sign":{"post":{"tags":["Auth"],"summary":"Performs user Login","description":"Method that performs user login to the system.","operationId":"signIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{},"application/xml":{},"application/yaml":{}}},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/auth/createUser":{"post":{"tags":["Auth"],"summary":"Creates the User","description":"Method to create a new user in the system.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}}},"required":true},"responses":{"200":{"description":"Bad Request"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/post/v1/{id}":{"get":{"tags":["Post"],"summary":"Find by one Post","description":"Find by one Post","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"delete":{"tags":["Post"],"summary":"Delete one Post","description":"Delete one Post","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/person/v1/{id}":{"get":{"tags":["Person"],"summary":"Find by one person","description":"Find by one person","operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"delete":{"tags":["Person"],"summary":"Delete one person","description":"Delete one person","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/like/v1/{id}":{"get":{"tags":["Like"],"summary":"Find by one Like","description":"Find by one Like","operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"delete":{"tags":["Like"],"summary":"Delete a one Like","description":"Delete a one Like","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/api/comment/v1/{id}":{"get":{"tags":["Comment"],"summary":"Find by one Comment","description":"Find by one Comment","operationId":"findById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"delete":{"tags":["Comment"],"summary":"Delete a Comment","description":"Delete a Comment","operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}}},"components":{"schemas":{"CommentDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDTO"},"postId":{"type":"integer","format":"int64"},"likes":{"type":"array","items":{"$ref":"#/components/schemas/LikeDTO"}},"parent":{"$ref":"#/components/schemas/ParentCommentDTO"},"_links":{"$ref":"#/components/schemas/Links"}}},"LikeDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserDTO"},"postId":{"type":"integer","format":"int64"},"commentId":{"type":"integer","format":"int64"},"_links":{"$ref":"#/components/schemas/Links"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"ParentCommentDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"}}},"PostDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"subTitle":{"type":"string"},"content":{"type":"string"},"date":{"type":"string"},"userId":{"type":"integer","format":"int64"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentDTO"}},"likes":{"type":"array","items":{"$ref":"#/components/schemas/LikeDTO"}},"_links":{"$ref":"#/components/schemas/Links"}}},"UserDTO":{"type":"object","properties":{"fullName":{"type":"string"},"username":{"type":"string"}}},"PersonDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"birthDate":{"type":"string","format":"date"},"phone":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"AccountCredentialsDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"fullname":{"type":"string"}}},"PostController":{"type":"object"},"PersonController":{"type":"object"},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}